2019/DataScotland/Demo/2 - Azure Kubernetes Service/AlwaysOn-AKS/Kubernetes/sample-deployment-script/requirements.txt
pyyaml

kubernetes
